web - extract logic into separate components to follow project code conventions
